+++ account-kit/react/src/hooks/useSolanaTransaction.ts
@@ -2,10 +2,9 @@
 
 import * as solanaNetwork from "../solanaNetwork.js";
 import { useMutation } from "@tanstack/react-query";
-import { SolanaSigner } from "@account-kit/signer";
+import { addSolanaSponsorship, SolanaSigner } from "@account-kit/signer";
 import type { BaseHookMutationArgs } from "../types.js";
 import {
-  Connection,
   PublicKey,
   SystemProgram,
   Transaction,
@@ -94,62 +93,6 @@ export type SolanaTransactionHookParams = {
    */
   mutation?: BaseHookMutationArgs<{ hash: string }, SolanaTransactionParams>;
 };
-
-async function addSponsorship(
-  instructions: TransactionInstruction[],
-  connection: Connection,
-  policyId: string,
-  address: string,
-): Promise<VersionedTransaction> {
-  const { blockhash } = await connection.getLatestBlockhash({
-    commitment: "finalized",
-  });
-  const message = new TransactionMessage({
-    // Right now the backend will rewrite this payer Key to the server's address
-    payerKey: new PublicKey(address),
-    recentBlockhash: blockhash,
-    instructions,
-  }).compileToV0Message();
-  const versionedTransaction = new VersionedTransaction(message);
-  const serializedTransaction = Buffer.from(
-    versionedTransaction.serialize(),
-  ).toString("base64");
-  const body = JSON.stringify({
-    id: crypto?.randomUUID() ?? Math.floor(Math.random() * 1000000),
-    jsonrpc: "2.0",
-    method: "alchemy_requestFeePayer",
-    params: [
-      {
-        policyId,
-        serializedTransaction,
-      },
-    ],
-  });
-  const options = {
-    method: "POST",
-    headers: {
-      accept: "application/json",
-      "content-type": "application/json",
-    },
-    body,
-  };
-
-  const response = await fetch(
-    // TODO: Use the connection??
-    connection.rpcEndpoint,
-    options,
-  );
-  const jsonResponse = await response.json();
-  if (!jsonResponse?.result?.serializedTransaction)
-    throw new Error(
-      `Response doesn't include the serializedTransaction ${JSON.stringify(
-        jsonResponse,
-      )}`,
-    );
-  return VersionedTransaction.deserialize(
-    Buffer.from(jsonResponse.result.serializedTransaction, "base64"),
-  );
-}
 
 /**
  * This is the hook that will be used to send a transaction.
@@ -259,7 +202,7 @@ export function useSolanaTransaction(
     "policyId" in opts ? opts.policyId : backupConnection?.policyId;
   const mapTransformInstructions: Record<string, TransformInstruction> = {
     async addSponsorship(instructions: TransactionInstruction[]) {
-      return await addSponsorship(
+      return await addSolanaSponsorship(
         instructions,
         connection || missing("connection"),
         policyId || missing("policyId"),
