+++ account-kit/signer/src/index.ts
@@ -21,3 +21,4 @@ export type * from "./solanaSigner.js";
 export type * from "./types.js";
 export { AlchemySignerStatus } from "./types.js";
 export { SolanaSigner } from "./solanaSigner.js";
+export { addSolanaSponsorship } from "./utils/solana.js";
